+++ Classes/DependencyInjection/Widgets/CampaignsRegistration.php
@@ -69,7 +69,7 @@ final class CampaignsRegistration extends AbstractRegistration
                         'column' => 'nb_visits',
                         'header' => Extension::LANGUAGE_PATH_DASHBOARD . ':visits',
                         'decorator' => new Reference(NumberDecorator::class),
-                        'classes' => 'text-right',
+                        'classes' => 'matomo-widgets__text-end',
                     ],
                 ],
             )