+++ Classes/DependencyInjection/Widgets/CountriesRegistration.php
@@ -80,7 +80,7 @@ final class CountriesRegistration extends AbstractRegistration
                         'column' => 'nb_visits',
                         'header' => Extension::LANGUAGE_PATH_DASHBOARD . ':visits',
                         'decorator' => new Reference(NumberDecorator::class),
-                        'classes' => 'text-right',
+                        'classes' => 'matomo-widgets__text-end',
                     ],
                 ],
             )