+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Proxy configuration is not taken into account (#46)
+
 ## [2.2.0] - 2023-05-21
 
 ### Added