+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Table content with numbers is right-aligned again in TYPO3 v12
+
 ## [2.2.1] - 2023-07-21
 
 ### Fixed