+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.0] - 2023-12-15
+
 ### Added
 - Widget "Most viewed pages" (#47)
 
@@ -172,7 +174,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release
 
 
-[Unreleased]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.2.1...HEAD
+[Unreleased]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.3.0...HEAD
+[2.3.0]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.2.1...v2.3.0
 [2.2.1]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.1.0...v2.1.1